Crate rbe

Regular Bag Expressions (rbe)

Provides an implementation of Regular Bag Expressions which are the expressions internally employed by the implementations of Shape Expressions More information about Regular Bag Expressions:

Complexity and Expressiveness of ShEx for RDF S. Staworko, I. Boneva, J. Labra, S. Hym, E. Prud’hommeaux, H. Solbrig

Re-exports

pub use crate::cardinality::*;

pub use crate::component::*;

pub use crate::deriv_n::*;

pub use crate::failures::*;

pub use crate::keys::*;

pub use crate::match_cond::*;

pub use crate::max::*;

pub use crate::min::*;

pub use crate::pending::*;

pub use crate::rbe1::*;

pub use crate::rbe1_matcher::*;

pub use crate::rbe_error::*;

pub use crate::rbe_schema::*;

pub use crate::rbe_table::*;

pub use crate::values::*;

pub use crate::bag::*;

Modules

bag

A set whose elements can be repeated. The set tracks how many times each element appears

candidate

cardinality

component

deriv_error

deriv_n

failures

keys

match_cond

max

min

pending

rbe

rbe1

rbe1_matcher

rbe_error

rbe_schema

rbe_table

values

Traits

Key

Ref

Value