rbe::rbe1

Enum Rbe

pub enum Rbe<K, V, R>
where
    K: Key,
    V: Value,
    R: Ref,
{
    Fail {
        error: RbeError<K, V, R>,
    },
    Empty,
    Symbol {
        key: K,
        cond: MatchCond<K, V, R>,
        card: Cardinality,
    },
    And {
        exprs: Vec<Rbe<K, V, R>>,
    },
    Or {
        exprs: Vec<Rbe<K, V, R>>,
    },
    Star {
        expr: Box<Rbe<K, V, R>>,
    },
    Plus {
        expr: Box<Rbe<K, V, R>>,
    },
    Repeat {
        expr: Box<Rbe<K, V, R>>,
        card: Cardinality,
    },
}

Variants

Fail

Fields

error: RbeError<K, V, R>

Empty

Symbol

Fields

key: K

cond: MatchCond<K, V, R>

card: Cardinality

And

Fields

exprs: Vec<Rbe<K, V, R>>

Or

Fields

exprs: Vec<Rbe<K, V, R>>

Star

Fields

expr: Box<Rbe<K, V, R>>

Plus

Fields

expr: Box<Rbe<K, V, R>>

Repeat

Fields

expr: Box<Rbe<K, V, R>>

card: Cardinality

Implementations

impl<K, V, R> Rbe<K, V, R>
where K: Key, V: Value, R: Ref,

pub fn empty() -> Rbe<K, V, R>

pub fn symbol(key: K, min: usize, max: Max) -> Rbe<K, V, R>

pub fn symbol_cond( key: K, cond: MatchCond<K, V, R>, min: Min, max: Max, ) -> Rbe<K, V, R>

pub fn or<I>(exprs: I) -> Rbe<K, V, R>
where I: IntoIterator<Item = Rbe<K, V, R>>,

pub fn and<I>(exprs: I) -> Rbe<K, V, R>
where I: IntoIterator<Item = Rbe<K, V, R>>,

pub fn opt(v: Rbe<K, V, R>) -> Rbe<K, V, R>

pub fn plus(expr: Rbe<K, V, R>) -> Rbe<K, V, R>

pub fn star(expr: Rbe<K, V, R>) -> Rbe<K, V, R>

pub fn repeat(expr: Rbe<K, V, R>, min: usize, max: Max) -> Rbe<K, V, R>

pub fn is_fail(&self) -> bool

pub fn symbols(&self) -> HashSet<K>

pub fn nullable(&self) -> bool

pub fn deriv( &self, symbol: &K, value: &V, n: usize, open: bool, controlled: &HashSet<K>, pending: &mut Pending<V, R>, ) -> Rbe<K, V, R>
where K: Eq + Hash + Clone,

Calculates the derivative of a rbe for a symbol with value open indicates if we allow extra symbols controlled contains the list of symbols controlled by the rbe that should not be assumed as extra symbols pending